apps/pi-controller/requirements.txt
luma.led_matrix>=1.7.1
luma.core>=2.4.2
websockets>=12.0
